Fix SP-MSE regression: remove inner bf16 autocast, use fp16 in score_fn
Transformer SP-MSE was 2x worse than SEDD baseline because the inner
bfloat16 autocast around the block stack (disabled in original SEDD)
was enabled here, and get_score_fn used bfloat16 instead of float16.
Restores parity with SEDD (transformer 0.0239, conv 0.0294 @ 100 steps).

+++ d3_dna/io.py
@@ -76,7 +76,7 @@ def get_score_fn(model, train=False, sampling=False):
 
     def score_fn(x, sigma, labels=None):
         device_type = 'cuda' if x.is_cuda else 'cpu'
-        with torch.amp.autocast(device_type, dtype=torch.bfloat16, enabled=(device_type == 'cuda')):
+        with torch.amp.autocast(device_type, dtype=torch.float16, enabled=(device_type == 'cuda')):
             sigma = sigma.reshape(-1)
             model_output = model_fn(x, sigma, labels)
             if isinstance(model_output, tuple):
